+++ template/__init__.py
@@ -10,7 +10,8 @@ call is made, and not when the module is loaded.
 
 There are three types of symbols in this package:
 
-1. model functions, such as `apec` and `TBabs`.
+1. model functions, such as `apec` and `TBabs`, and `tableModel` for
+   table models.
 2. routines that set or get values such as the abundance
    table (`abundance`), cross-section table (`cross_sections`),
    cosmology (`cosmology`), and the chatter level (`chatter`).
@@ -75,6 +76,53 @@ which would keep the original model values.
 >>> x.kdblur(energies=egrid, pars=pars_kdblur, model=ymodel)
 >>> plt.plot(emid, ymodel, alpha=0.8, label='Convolved')
 >>> plt.legend()
+
+XSPEC table models [TableModel]_ are fun to work with, as you
+
+1. need to read in the file to find out information on the model -
+   such as whether it's atable or mtable (but unfortunately there is
+   no header keyword to determine if it is an etable) - and the
+   parameter names, values, and ranges.
+
+2. use the file name when evaluating the model along with some of
+   this metadata.
+
+At the moment this module only supports the second part - calling the
+models - and it is left to the user to find the other information out.
+
+In this example the ``RCS.mod`` table model, which has three
+parameters, does not add a redshift parameter, and is an "atable"
+model (i.e. additive):
+
+    % dmlist "RCS.mod[cols name, initial]" data,clean
+    #  NAME           INITIAL
+     tau                             1.0
+     beta               0.10000000149012
+     T                  0.10000000149012
+    % dmkeypar xspec-tablemodel-RCS.mod"[primary]" redshift echo+
+    0
+    % dmkeypar "xspec-tablemodel-RCS.mod[primary]" addmodel echo+
+    1
+
+This can then be used with `tableModel` in a similar manner to the
+other models, apart from requiring `table` and `table_type` arguments:
+
+>>> infile = 'RCS.mod'
+>>> pars = [1, 0.1, 0.1]
+>>> egrid = np.arange(0.1, 10, 0.01)
+>>> y = x.tableModel(table=infile, table_type="add", energies=egrid, pars=pars)
+
+Note that it is very easy to make the table model code crash the
+system, such as by sending in not enough parameters or settnig a
+parameter outside its hard limits:
+
+>>> x.tableModel(infile, "add", pars=[1, 2], energies=egrid)
+Segmentation fault (core dumped)
+
+References
+----------
+
+.. [TableModel] https://heasarc.gsfc.nasa.gov/docs/heasarc/ofwg/docs/general/ogip_92_009/ogip_92_009.html
 
 """